Combine all e2e tests into a single file
- Combined all e2e tests into extension.spec.ts
- Organized tests into logical groups
- Made test patterns consistent
- Added common setup and teardown hooks
- Removed redundant test files
